database

+++ delete_sentsms.php
@@ -1,10 +1,18 @@
 <?php 
 
-include "database.php";
-$sql = " DELETE FROM sentsms WHERE id = $id";
-$result = mysqli_query($conn,$sql) or die(mysqli_error($conn));
 
-$_SERVER['HTTP_REFERER'];
+include "database.php";
+
+if ($_GET['id']) {
+
+	$id =$_GET['id'];
+	$sql = "DELETE FROM sentsms WHERE id = $id";
+	$result = mysqli_query($conn,$sql) or die(mysqli_error($conn));
+
+	
+	header("sentsms.php");
+}
+
